Tidied the tests a little, plus updated readme and added selenium grid docker compose
